# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the public pool api from stratum.farm",
        "title": "1oop Pool API",
        "termsOfService": "https://stratum.farm/terms/",
        "contact": {
            "name": "1oop Support",
            "email": "pool@stratum.farm"
        },
        "version": "1.0"
    },
    "host": "152.228.229.130:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/pools": {
            "get": {
                "description": "Get a list of all available pools",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get all pools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.PoolsRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/pools/{pool_id}": {
            "get": {
                "description": "Get a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get a pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Range in blocks to fetch the average effort from (default=50)",
                        "name": "effortRange",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PoolExtendedRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/pools/{pool_id}/blocks": {
            "get": {
                "description": "Get a list of blocks from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get a list of blocks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page (default=0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PageSize (default=15)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BlocksRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pools/{pool_id}/payments": {
            "get": {
                "description": "Get a list of payments from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get a list of payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page (default=0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PageSize (default=15)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PaymentsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/pools/{pool_id}/performance": {
            "get": {
                "description": "Get a list of performance samples from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get a list of performance samples",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sample interval (default=Hour)",
                        "name": "i",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sample range (default=Day)",
                        "name": "r",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PoolPerformanceRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/pools/{pool_id}/topminers": {
            "get": {
                "description": "Get the top miners from a specific pool",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get the top miners from a pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Range in hours to fetch the top miners from (default=1)",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TopMinersRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Block": {
            "type": "object",
            "properties": {
                "blockHeight": {
                    "type": "integer"
                },
                "confirmationProgress": {
                    "type": "number"
                },
                "created": {
                    "type": "string"
                },
                "effort": {
                    "type": "number"
                },
                "hash": {
                    "type": "string"
                },
                "infoLink": {
                    "type": "string"
                },
                "miner": {
                    "type": "string"
                },
                "networkDifficulty": {
                    "type": "number"
                },
                "poolId": {
                    "type": "string"
                },
                "reward": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionConfirmationData": {
                    "type": "string"
                }
            }
        },
        "api.BlocksRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Block"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Payment": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "addressInfoLink": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "coin": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "transactionConfirmationData": {
                    "type": "string"
                },
                "transactionInfoLink": {
                    "type": "string"
                }
            }
        },
        "api.PaymentsRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Payment"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Pool": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "blockHeight": {
                    "type": "integer"
                },
                "coin": {
                    "type": "string"
                },
                "fee": {
                    "type": "number"
                },
                "feeType": {
                    "type": "string"
                },
                "hashrate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "miners": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "networkDifficulty": {
                    "type": "number"
                },
                "networkHashrate": {
                    "type": "number"
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/api.Price"
                    }
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "api.PoolEndpoint": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "type": "number"
                },
                "tls": {
                    "type": "boolean"
                },
                "tlsAuto": {
                    "type": "boolean"
                },
                "varDiff": {
                    "type": "boolean"
                }
            }
        },
        "api.PoolExtended": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "algorithm": {
                    "type": "string"
                },
                "averageEffort": {
                    "type": "number"
                },
                "blockHeight": {
                    "type": "integer"
                },
                "coin": {
                    "type": "string"
                },
                "effort": {
                    "type": "number"
                },
                "fee": {
                    "type": "number"
                },
                "feeType": {
                    "type": "string"
                },
                "hashrate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lastBlockFoundTime": {
                    "type": "string"
                },
                "minPayout": {
                    "type": "number"
                },
                "miners": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "networkDifficulty": {
                    "type": "number"
                },
                "networkHashrate": {
                    "type": "number"
                },
                "ports": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/api.PoolEndpoint"
                    }
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/api.Price"
                    }
                },
                "totalBlocksFound": {
                    "type": "integer"
                },
                "totalPayments": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "api.PoolExtendedRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "$ref": "#/definitions/api.PoolExtended"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.PoolPerformance": {
            "type": "object",
            "properties": {
                "connectedMiners": {
                    "type": "integer"
                },
                "created": {
                    "type": "string"
                },
                "networkDifficulty": {
                    "type": "number"
                },
                "networkHashrate": {
                    "type": "number"
                },
                "poolHashrate": {
                    "type": "number"
                }
            }
        },
        "api.PoolPerformanceRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.PoolPerformance"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.PoolsRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Pool"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.Price": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "priceChangePercentage24H": {
                    "type": "number"
                }
            }
        },
...
        "api.TopMiner": {
            "type": "object",
            "properties": {
                "hashrate": {
                    "type": "number"
                },
                "joined": {
                    "type": "string"
                },
                "miner": {
                    "type": "string"
                },
                "totalPaid": {
                    "type": "number"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "api.TopMinersRes": {
            "type": "object",
            "properties": {
                "pageCount": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TopMiner"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}